# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MimiiHub e-commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with optional filters
      parameters:
        - { name: categoryId, in: query, schema: { type: number } }
        - { name: subcategoryId, in: query, schema: { type: number } }
        - { name: featured, in: query, schema: { type: boolean } }
        - { name: newArrival, in: query, schema: { type: boolean } }
        - { name: bestSeller, in: query, schema: { type: boolean } }
        - { name: search, in: query, schema: { type: string } }
        - { name: visible, in: query, schema: { type: boolean } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product
      parameters:
        - { name: id, in: path, required: true, schema: { type: number } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - { name: id, in: path, required: true, schema: { type: number } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - { name: id, in: path, required: true, schema: { type: number } }
      responses:
        "204":
          description: Deleted
  /products/{id}/related:
    get:
      operationId: getRelatedProducts
      tags: [products]
      summary: Get related products (same category)
      parameters:
        - { name: id, in: path, required: true, schema: { type: number } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  # ── ORDERS ────────────────────────────────────────────────────
...
  /settings/store:
    get:
      operationId: getStoreSettings
      tags: [settings]
      summary: Get store settings
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreSettings"
    patch:
      operationId: updateStoreSettings
      tags: [settings]
      summary: Update store settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoreSettingsUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreSettings"
  /settings/homepage:
    get:
      operationId: getHomepageSettings
      tags: [settings]
      summary: Get homepage settings (banners, trust bar, collections)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HomepageSettings"
    patch:
      operationId: updateHomepageSettings
      tags: [settings]
      summary: Update homepage settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HomepageSettingsUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HomepageSettings"
  /settings/payment:
    get:
      operationId: getPaymentSettings
      tags: [settings]
      summary: Get payment settings
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentSettings"
    patch:
      operationId: updatePaymentSettings
      tags: [settings]
      summary: Update payment settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentSettingsUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentSettings"
components:
  schemas:
...
    Category:
      type: object
      required: [id, name, slug, createdAt]
      properties:
        id:          { type: number }
        name:        { type: string }
        slug:        { type: string }
        description: { type: ["string", "null"] }
        image:       { type: ["string", "null"] }
        createdAt:   { type: string, format: date-time }
        subcategories:
          type: array
          items:
            $ref: "#/components/schemas/Subcategory"
...
    Subcategory:
      type: object
      required: [id, categoryId, name, slug, createdAt]
      properties:
        id:         { type: number }
        categoryId: { type: number }
        name:       { type: string }
        slug:       { type: string }
        createdAt:  { type: string, format: date-time }
...
    ProductSpec:
      type: object
      properties:
        capacity:   { type: ["string", "null"] }
        weight:     { type: ["string", "null"] }
        dimensions: { type: ["string", "null"] }
        size:       { type: ["string", "null"] }
        color:      { type: ["string", "null"] }
        material:   { type: ["string", "null"] }
    Product:
      type: object
      required: [id, name, slug, price, images, visible, inStock, featured, newArrival, bestSeller, createdAt]
      properties:
        id:            { type: number }
        name:          { type: string }
        slug:          { type: string }
        description:   { type: ["string", "null"] }
        price:         { type: number }
        discountPct:   { type: ["number", "null"] }
        discountedPrice: { type: ["number", "null"] }
        images:
          type: array
          items: { type: string }
        coverImage:    { type: ["string", "null"] }
        categoryId:    { type: ["integer", "null"] }
        subcategoryId: { type: ["integer", "null"] }
        category:      { $ref: "#/components/schemas/Category", nullable: true }
        subcategory:   { $ref: "#/components/schemas/Subcategory", nullable: true }
        stockQty:      { type: number }
        inStock:       { type: boolean }
        visible:       { type: boolean }
        featured:      { type: boolean }
        newArrival:    { type: boolean }
        bestSeller:    { type: boolean }
        specs:         { $ref: "#/components/schemas/ProductSpec" }
        createdAt:     { type: string, format: date-time }
    ProductInput:
      type: object
      required: [name, slug, price, images]
      properties:
        name:          { type: string }
        slug:          { type: string }
        description:   { type: string }
        price:         { type: number }
        discountPct:   { type: number }
        images:        { type: array, items: { type: string } }
        coverImage:    { type: string }
        categoryId:    { type: number }
        subcategoryId: { type: number }
        stockQty:      { type: number }
        inStock:       { type: boolean }
        visible:       { type: boolean }
        featured:      { type: boolean }
        newArrival:    { type: boolean }
        bestSeller:    { type: boolean }
        specs:         { $ref: "#/components/schemas/ProductSpec" }
    ProductUpdate:
      type: object
      properties:
        name:          { type: string }
        slug:          { type: string }
        description:   { type: string }
        price:         { type: number }
        discountPct:   { type: ["number", "null"] }
        images:        { type: array, items: { type: string } }
        coverImage:    { type: ["string", "null"] }
        categoryId:    { type: ["integer", "null"] }
        subcategoryId: { type: ["integer", "null"] }
        stockQty:      { type: number }
        inStock:       { type: boolean }
        visible:       { type: boolean }
        featured:      { type: boolean }
        newArrival:    { type: boolean }
        bestSeller:    { type: boolean }
        specs:         { $ref: "#/components/schemas/ProductSpec" }
...
    TrustItem:
      type: object
      required: [id, title, subtitle, icon]
      properties:
        id:       { type: string }
        title:    { type: string }
        subtitle: { type: string }
        icon:     { type: string }
    HeroBanner:
      type: object
      required: [id, title, subtitle, image]
      properties:
        id:         { type: string }
        title:      { type: string }
        subtitle:   { type: string }
        image:      { type: string }
        buttonText: { type: ["string", "null"] }
        buttonLink: { type: ["string", "null"] }
    FeaturedCollection:
      type: object
      required: [id, title, description, image, link]
      properties:
        id:          { type: string }
        title:       { type: string }
        description: { type: string }
        image:       { type: string }
        link:        { type: string }
    HomepageSettings:
      type: object
      required: [heroBanners, trustItems, featuredCollections]
      properties:
        heroBanners:
          type: array
          items:
            $ref: "#/components/schemas/HeroBanner"
        trustItems:
          type: array
          items:
            $ref: "#/components/schemas/TrustItem"
        featuredCollections:
          type: array
          items:
            $ref: "#/components/schemas/FeaturedCollection"
    HomepageSettingsUpdate:
      type: object
      properties:
        heroBanners:
          type: array
          items:
            $ref: "#/components/schemas/HeroBanner"
        trustItems:
          type: array
          items:
            $ref: "#/components/schemas/TrustItem"
        featuredCollections:
          type: array
          items:
            $ref: "#/components/schemas/FeaturedCollection"
    SocialLinks:
      type: object
      properties:
        instagram: { type: ["string", "null"] }
        facebook:  { type: ["string", "null"] }
        twitter:   { type: ["string", "null"] }
        tiktok:    { type: ["string", "null"] }
    DeliverySettings:
      type: object
      properties:
        freeDeliveryThreshold: { type: ["number", "null"] }
        deliveryFee:           { type: number }
        deliveryNote:          { type: ["string", "null"] }
    StoreSettings:
      type: object
      required: [storeName, email]
      properties:
        storeName:        { type: string }
        logo:             { type: ["string", "null"] }
        contactPhone:     { type: ["string", "null"] }
        whatsapp:         { type: ["string", "null"] }
        email:            { type: string }
        socialLinks:      { $ref: "#/components/schemas/SocialLinks" }
        deliverySettings: { $ref: "#/components/schemas/DeliverySettings" }
    StoreSettingsUpdate:
      type: object
      properties:
        storeName:        { type: string }
        logo:             { type: ["string", "null"] }
        contactPhone:     { type: ["string", "null"] }
        whatsapp:         { type: ["string", "null"] }
        email:            { type: string }
        socialLinks:      { $ref: "#/components/schemas/SocialLinks" }
        deliverySettings: { $ref: "#/components/schemas/DeliverySettings" }
    PaymentSettings:
      type: object
      required: [flutterwaveEnabled, payOnDeliveryEnabled]
      properties:
        flutterwaveEnabled:    { type: boolean }
        payOnDeliveryEnabled:  { type: boolean }
        flutterwavePublicKey:  { type: ["string", "null"] }
    PaymentSettingsUpdate:
      type: object
      properties:
        flutterwaveEnabled:   { type: boolean }
        payOnDeliveryEnabled: { type: boolean }