# lib/api-zod/src/generated/types/featuredCollection.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * MimiiHub e-commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface FeaturedCollection {
  id: string;
  title: string;
  description: string;
  image: string;
  link: string;
}