# lib/api-zod/src/generated/types/heroBanner.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * MimiiHub e-commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface HeroBanner {
  id: string;
  title: string;
  subtitle: string;
  image: string;
  /** @nullable */
  buttonText?: string | null;
  /** @nullable */
  buttonLink?: string | null;
}