# lib/api-zod/src/generated/types/paymentSettingsUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * MimiiHub e-commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface PaymentSettingsUpdate {
  flutterwaveEnabled?: boolean;
  payOnDeliveryEnabled?: boolean;
}