# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * MimiiHub e-commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { Category } from './category';
import type { ProductSpec } from './productSpec';
import type { Subcategory } from './subcategory';

export interface Product {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  discountPct?: number | null;
  /** @nullable */
  discountedPrice?: number | null;
  images: string[];
  /** @nullable */
  coverImage?: string | null;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  subcategoryId?: number | null;
  category?: Category | null;
  subcategory?: Subcategory | null;
  stockQty?: number;
  inStock: boolean;
  visible: boolean;
  featured: boolean;
  newArrival: boolean;
  bestSeller: boolean;
  specs?: ProductSpec;
  createdAt: Date;
}