# lib/api-zod/src/generated/types/productSpec.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * MimiiHub e-commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductSpec {
  /** @nullable */
  capacity?: string | null;
  /** @nullable */
  weight?: string | null;
  /** @nullable */
  dimensions?: string | null;
  /** @nullable */
  size?: string | null;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  material?: string | null;
}