# lib/api-zod/src/generated/types/productUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * MimiiHub e-commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductSpec } from './productSpec';

export interface ProductUpdate {
  name?: string;
  slug?: string;
  description?: string;
  price?: number;
  /** @nullable */
  discountPct?: number | null;
  images?: string[];
  /** @nullable */
  coverImage?: string | null;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  subcategoryId?: number | null;
  stockQty?: number;
  inStock?: boolean;
  visible?: boolean;
  featured?: boolean;
  newArrival?: boolean;
  bestSeller?: boolean;
  specs?: ProductSpec;
}