# lib/api-zod/src/generated/types/storeSettingsUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * MimiiHub e-commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { DeliverySettings } from './deliverySettings';
import type { SocialLinks } from './socialLinks';

export interface StoreSettingsUpdate {
  storeName?: string;
  /** @nullable */
  logo?: string | null;
  /** @nullable */
  contactPhone?: string | null;
  /** @nullable */
  whatsapp?: string | null;
  email?: string;
  socialLinks?: SocialLinks;
  deliverySettings?: DeliverySettings;
}